+++ packages/spacecat-shared-project-engine-client/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [@adobe/spacecat-shared-project-engine-client-v1.8.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-project-engine-client-v1.7.1...@adobe/spacecat-shared-project-engine-client-v1.8.0) (2026-07-03)
+
+### Features
+
+* **project-engine-client:** add GET /v1/url/resolve to the vendored client + mock ([#1771](https://github.com/adobe/spacecat-shared/issues/1771)) ([e8f22b7](https://github.com/adobe/spacecat-shared/commit/e8f22b75db738a6c10f48796a521d55a8b79f2e3)), closes [#737](https://github.com/adobe/spacecat-shared/issues/737) [serenity-docs#25](https://github.com/adobe/serenity-docs/issues/25) [serenity-docs#25](https://github.com/adobe/serenity-docs/issues/25)
+
+## [@adobe/spacecat-shared-project-engine-client-v1.7.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-project-engine-client-v1.7.0...@adobe/spacecat-shared-project-engine-client-v1.7.1) (2026-07-02)
+
+### Bug Fixes
+
+* **project-engine-client:** retype aio-create-prompt-v2's 200 response (CR14) ([#1765](https://github.com/adobe/spacecat-shared/issues/1765)) ([33315e1](https://github.com/adobe/spacecat-shared/commit/33315e1bcd3d7d789aaec9e1a72c12e69bfa7799)), closes [#1764](https://github.com/adobe/spacecat-shared/issues/1764) [#1764](https://github.com/adobe/spacecat-shared/issues/1764)
+
 ## [@adobe/spacecat-shared-project-engine-client-v1.7.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-project-engine-client-v1.6.0...@adobe/spacecat-shared-project-engine-client-v1.7.0) (2026-07-02)
 
 ### Features